Download Dictionaries

Downloads the specified dictionaries from osf.io/y6g5b.

Usage

download.dict(dict = "lusi", check.md5 = TRUE, mode = "wb",
  dir = getOption("lingmatch.dict.dir"), overwrite = FALSE)

Arguments

dict

One or more names of dictionaries to download, or 'all' for all available. See osf.io/y6g5b/wiki for more information, and a list of available dictionaries.

check.md5

Logical; if TRUE (default), retrieves the MD5 checksum from OSF, and compares it with that calculated from the downloaded file to check its integrity.

mode

A character specifying the file write mode; default is 'wb'. See download.file.

dir

Directory in which to save the dictionary;
default is getOption('lingmatch.dict.dir').
This must be specified, or the option must be set – use lma_initdirs to initialize a directory.

overwrite

Logical; if TRUE, will replace existing files.

Value

Path to the downloaded dictionary, or a list of such if multiple were downloaded.

See also

Examples

if (FALSE) { # \dontrun{

download.dict("lusi", dir = "~/Dictionaries")
} # }